Modified the reg

+++ src/utils.ts
@@ -22,9 +22,9 @@ export const getTransformValue = (
     fontViewportUnit,
   } = options
   const unit = ~prop.indexOf('font') ? fontViewportUnit : viewportUnit
-  if (Math.abs(parseInt(string)) === 0) return string
-  if (Math.abs(parseInt(string)) <= minPixelValue) return string
-  const value = parseInt(string) / viewportWidth
+  if (Math.abs(parseFloat(string)) === 0) return string
+  if (Math.abs(parseFloat(string)) <= minPixelValue) return string
+  const value = parseFloat(string) / viewportWidth
   let result = `${toFixed(value * 100, unitPrecision)}${unit}`
 
   if (maxViewportWidth) {
